CP-24350: expose sharable flag to SMAPIv3 backends
The HA statefile and redo log needs to be attachable from multiple hosts, which
used to be done as a side-effect of sm-config:type=raw in SMAPIv1.
For SMAPIv3 we'll introduce a sharable flag in the vdi creation phase to make
the semantics clearer, because  for GFS2 what matters is not whether the file
format is raw or not, but whether it can be attached from multiple hosts.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>